refactor: Actualizar servicio de eliminación de tipos de equipo
Actualizar el servicio de eliminación de tipos de equipo para incluir la validación de que no se pueden eliminar los tipos de equipo que tienen equipos asociados.

+++ src/features/equipment-types/equipment-types.service.ts
@@ -1,4 +1,5 @@
 import EquipmentType from "../../models/equipment-type.model";
+import Equipment from "../../models/equipment.model";
 import {
   IEquipmentTypeCreationAttributes,
   IEquipmentTypeUpdateAttributes,
@@ -46,8 +47,25 @@ export class EquipmentTypesService {
   }
 
   async deleteEquipmentType(id: string) {
+
+    const foundEquipmentType = await this.equipmentTypesModel.findByPk(id,
+      {
+        include: [{ model: Equipment, as: "equipments" }],
+      }
+    )
+
+    if (!foundEquipmentType) {
+      throw new Error("Tipo de equipo no encontrado");
+           
+    }
+
+    if (foundEquipmentType.equipments.length) {
+      throw new Error("No se puede eliminar el tipo de equipo porque tiene equipos asociados.");
+    }
+
+
     return await this.equipmentTypesModel.destroy({
-      where: { equipment_type_id: id },
+      where: { equipment_type_id: foundEquipmentType.equipment_type_id },
     });
   }
 }
